requirements.txt
flet
telethon
google-generativeai